changed variable from uint256 to uint64 (max len of a document)

+++ document.sol
@@ -7,7 +7,7 @@ contract Document is DocumentAbstract {
     dataLen = len;
     collection = _c;
 
-    for (uint256 i = 0; i < len; i++) {
+    for (uint64 i = 0; i < len; i++) {
       data.push(_data[i]);
     }
   }
